Changes towards resolving WIP status - Added copyright and license to files - Split and updated READMEs - Refactored writer_benchmarking, splitting classes into separate files - added ctrl-c handling to writer_benchmarking, in script and in producers
