+++ rosbag2_performance/rosbag2_benchmarking/plotting/plot_logging_data.py
@@ -1,3 +1,17 @@
+# Copyright 2020, Robotec.ai sp. z o.o.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http:#www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os, pathlib, re
 import matplotlib.pyplot as plt
 import numpy as np
@@ -27,7 +41,7 @@ with open(str(raport_path), "r") as raport1:
     first_line = raport1.readline()
     start_time = float(re.search('([0-9]*\.[0-9]*).*', first_line).group(1))
     print(start_time)
-    
+
     lines = raport1.readlines()
     for line in [first_line] + lines:
