+++ rosbag2_performance/rosbag2_benchmarking/rosbag2_benchmarking/voyager.py
@@ -1,3 +1,17 @@
+# Copyright 2020, Robotec.ai sp. z o.o.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import shutil
 import pathlib, yaml, csv
 import numpy as np
@@ -42,7 +56,7 @@ class VoyagerCase(Node):
                 self.config = yaml.load(config_file)
         else:
             raise RuntimeError("{} is not correct yaml config file.".format(path))
-        
+
         self.generate_voyager_results()
 
     def generate_voyager_results(self):
@@ -74,7 +88,7 @@ class VoyagerCase(Node):
         print("1000KB \t",table[12:16])
         print()
 
-        
+
         for row in range(0, 4):
             for col in range(0, 4):
                 val = "n/a" if row*4+col >= len(table) else table[row*4+col]
