+++ rosbag2_performance/rosbag2_performance_workers/README.md
@@ -1,0 +1,7 @@
+# Rosbag2 performance workers
+
+Publishers of random images, point clouds and byte messages, including scaling up and parametrization. Workers random image and pc2 generation part was based on Martin Idel [code](https://github.com/Martin-Idel/rosbag2/tree/performance_testing).
+
+## Usage
+
+These publishers are not ran manually, but used by rosbag2_benchmarking package.
