+++ rosbag2_performance/rosbag2_performance_workers/include/worker.hpp
@@ -1,3 +1,20 @@
+// Copyright 2020, Robotec.ai sp. z o.o.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef ROSBAG2_PERFORMANCE_WORKERS_HPP
+#define ROSBAG2_PERFORMANCE_WORKERS_HPP
+
 #include <chrono>
 #include <string>
 #include <vector>
@@ -105,3 +122,5 @@ private:
     rclcpp::TimerBase::SharedPtr delay_timer;
     std::vector<publisherPtr> publishers;
 };
+
+#endif //ROSBAG2_PERFORMANCE_WORKERS_HPP
