+++ rosbag2_performance/writer_benchmarking/include/byte_producer.hpp
@@ -1,0 +1,77 @@
+// Copyright 2020, Robotec.ai sp. z o.o.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef ROSBAG2_PERFORMANCE__BYTE_PRODUCER_HPP
+#define ROSBAG2_PERFORMANCE__BYTE_PRODUCER_HPP
+
+#include <chrono>
+#include <memory>
+#include <thread>
+#include "message_queue.hpp"
+#include "rclcpp/rclcpp.hpp" //ok()
+
+struct ProducerConfig
+{
+  unsigned int frequency;
+  unsigned int max_count;
+  unsigned int message_size;
+};
+
+class ByteProducer
+{
+public:
+  ByteProducer(const ProducerConfig &config, std::shared_ptr<ByteMessageQueue> queue) : mConfiguration(config), mQueue(queue)
+  {
+    generateRandomMessage();
+    mMsSleepTime = mConfiguration.frequency == 0 ? 1 : 1000/mConfiguration.frequency;
+  }
+
+  void run()
+  {
+    for (unsigned int i = 0; i < mConfiguration.max_count; ++i)
+    {
+      if (!rclcpp::ok())
+        break;
+      mQueue->push(mMessage);
+      std::this_thread::sleep_for(std::chrono::milliseconds(mMsSleepTime));
+    }
+    mQueue->setComplete();
+  }
+
+private:
+  std::vector<uint8_t> randomByteArrayData(size_t size)
+  {
+    std::vector<uint8_t> byte(size, 0);
+    for (size_t i = 0; i < size; ++i) {
+      byte[i] = std::rand() % 255;
+    }
+    return byte;
+  }
+
+  //Reuses the same random message to remove generation time from benchmarks
+  void generateRandomMessage()
+  {
+    mMessage.data = randomByteArrayData(mConfiguration.message_size);
+  }
+
+  std_msgs::msg::ByteMultiArray mMessage;
+  std::mutex mMutex;
+
+  ProducerConfig mConfiguration;
+  std::shared_ptr<ByteMessageQueue> mQueue;
+
+  unsigned int mMsSleepTime;
+};
+
+#endif //ROSBAG2_PERFORMANCE__BYTE_PRODUCER_HPP
