+++ rosbag2_performance/writer_benchmarking/scripts/benchmark.sh
@@ -1,3 +1,10 @@
+trap ctrlc SIGINT
+ctrlc_sent=0
+
+function ctrlc()
+{
+  let ctrlc_sent=1
+}
 
 test_dir=/tmp/rosbag2_test
 echo ${test_dir}
@@ -31,6 +38,10 @@ do
         echo "Results will be written to file: ${outfile}"
         ros2 run writer_benchmarking writer_benchmark --ros-args -p frequency:=${freq} -p size:=${sz} -p instances:=${inst} -p max_cache_size:=${cache} -p db_folder:=${db_path} 2> ${outfile}
         rm -fr ${db_path}/bag*
+        if [[ $ctrlc_sent -eq 1 ]]; then
+          echo -e "\e[31mQuitting prematurely due to Ctrl-C - some results aren't saved and some won't be reliable\e[0m]"
+          exit
+        fi
       done
     done
   done
